Move test to its own file because it was getting too long
